backend/requirements.txt
yfinance 
langchain_pinecone 
openai 
python-dotenv 
langchain-community
sentence_transformers
starlette
